+++ spec/dummy/client/entries/app-kitten.js
@@ -338,6 +338,7 @@ import {
 import {
   KarlCarouselProjectCard,
 } from 'kitten/karl/carousel/carousel'
+import { KarlSimpleCarousel } from 'kitten/karl/carousel/simple-carousel'
 
 ReactOnRails.register({
   // Box
@@ -615,4 +616,5 @@ ReactOnRails.register({
 
   // Carousel
   KarlCarouselProjectCard,
+  KarlSimpleCarousel,
 })